src/utils.py
from databricks.sdk import WorkspaceClient

def retrieve_workspace_url():
    # Here, you could implement logic to retrieve the workspace URL securely
    return 'workspace_url'

def retrieve_token():
    # Here, you could implement logic to retrieve the token securely
    return 'token'

def create_workspace_client():
    try:
        host = retrieve_workspace_url()
        token = retrieve_token()
        w = WorkspaceClient(host=host, token=token)
        return w
    except Exception as e:
        raise RuntimeError(f"Error creating workspace client: {e}")